Extensions: restore per-extension permission controls and hide the Install button until package installs ship
